requirements.txt
gunicorn==20.0.4
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
colorclass==2.2.0
coverage==5.2.1
distlib==0.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-environ==0.4.5
django-globals==0.3.2
django-test-without-migrations==0.6
django-webpack-loader==0.7.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
filelock==3.0.12
future==0.18.2
idna==2.10
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.5
openpyxl-templates==0.2.5
packaging==20.4
psycopg2==2.8.6
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
terminaltables==3.1.0
urllib3==1.25.10
wrapt==1.11.1
pyyaml==5.3.1